divide now returns float

+++ src/main/java/com/example/springdocker/MyMathCalc/MyMathCalc.java
@@ -10,7 +10,7 @@ public class MyMathCalc {
         return a * b;
     }
 
-    public int divide(int a, int b) {
+    public float divide(float a, float b) {
         return a / b;
     }
 }
